requirements.txt
Pillow>=10.2.0
flup6>=1.1.1
